add cancel order cancel rp

+++ apps/user/views.py
@@ -12,7 +12,7 @@ from utils.winxin import WeiXinAPI
 from market.models import GroupBuyGoods, GroupBuy
 from iuser.models import GenericOrder, UserProfile
 from models import UnpackRedPacketsLog
-from  MySQLdb import escape_string
+from MySQLdb import escape_string
 
 from iuser.Authentication import Authentication
 
@@ -282,16 +282,31 @@ class ConsumerOrderView(APIView):
 
         return Response(format_body(1, 'Success', {'id': group_buy_goods.group_buy_id, 'award_red_packets': award_red_packets}))
 
-    @raise_general_exception
     @Authentication.token_required
+    @raise_general_exception
     def delete(self, request):
         # 删除订单
-        order_goods = GenericOrder.objects.get(pk=request.GET['order_id'])
+        order_id = request.data['order_id'] if request.data.has_key('order_id') else request.GET['order_id']
+        order_goods = GenericOrder.objects.get(pk=order_id)
+        if not self.delete.user_id == order_goods.user_id:
+            return Response(format_body(22, 'Fail', 'xiajibagao'))
         order_goods.delete()
 
         #更新库存
         order_goods.goods.stock += order_goods.quantity
         order_goods.goods.save()
+
+        #该商品是否有红包, 有则失效
+        if int(order_goods.goods.group_buy.award_red_packets):
+            remain_order = GenericOrder.objects.filter(
+                user_id = self.delete.user_id,
+                goods__group_buy_id = order_goods.goods.group_buy_id
+            ).count()
+            if not remain_order:
+                UnpackRedPacketsLog.objects.filter(
+                    receiver = self.delete.user_id,
+                    group_buying_id = order_goods.goods.group_buy_id
+                ).update(is_failure = 2)
 
         return Response(format_body(1, 'Success', ''))
